Rename tests and compare with v2
